[upd] Added the possibility to process or delete only a subset of all the opened images
Each image now has a checkbox, that interacts with a unique master checkbox in the main window. When the user clicks on the Process Images button, only the images whose checkboxes are checked are processed.

The main window now also has a master delete button. When clicking on this button, all the images that have their checkboxes checked will be deleted.